Fix "Class path contains multiple SLF4J bindings" during runtime attachment
