+++ runtime-attach-distro/README.md
@@ -1,0 +1,5 @@
+# Runtime attachment distro
+
+This project is dedicated to providers of the [OpenTelemetry Java agent](https://github.com/open-telemetry/opentelemetry-java-instrumentation). It helps them to build a dependency to runtime attach their agent jar file.
+
+You can find [here](../runtime-attach/README.md) the project of the distribution published by OpenTelemetry.
