+++ runtime-attach-distro/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+  id("otel.java-conventions")
+  id("otel.publish-conventions")
+}
+
+description = "To help in create an OpenTelemetry distro able to runtime attach an OpenTelemetry Java Instrumentation agent"
+
+dependencies {
+  implementation("net.bytebuddy:byte-buddy-agent:1.11.18")
+
+  // Used by byte-buddy but not brought in as a transitive dependency.
+  compileOnly("com.google.code.findbugs:annotations")
+}
