+++ runtime-attach/build.gradle.kts
@@ -3,11 +3,13 @@ plugins {
   id("otel.publish-conventions")
 }
 
-description = "Utility to attach OpenTelemetry Java Instrumentation agent from classpath"
+description = "To runtime attach the OpenTelemetry Java Instrumentation agent"
+
+val agent: Configuration by configurations.creating
 
 dependencies {
-  compileOnly("io.opentelemetry.javaagent:opentelemetry-javaagent:1.6.0")
-  implementation("net.bytebuddy:byte-buddy-agent:1.11.18")
+  implementation(project(":runtime-attach-distro"))
+  agent("io.opentelemetry.javaagent:opentelemetry-javaagent:1.15.0")
 
   // Used by byte-buddy but not brought in as a transitive dependency.
   compileOnly("com.google.code.findbugs:annotations")
@@ -22,4 +24,14 @@ dependencies {
 tasks.test {
   useJUnitPlatform()
   setForkEvery(1) // One JVM by test class to avoid a test class launching a runtime attachment influences the behavior of another test class
+}
+
+tasks {
+  jar {
+    inputs.files(agent)
+    from({
+      agent.singleFile
+    })
+    rename { "otel-agent.jar" }
+  }
 }
