+++ settings.gradle.kts
@@ -47,6 +47,7 @@ include(":micrometer-meter-provider")
 include(":jmx-metrics")
 include(":maven-extension")
 include(":runtime-attach")
+include(":runtime-attach-distro")
 include(":samplers")
 include(":static-instrumenter:agent-instrumenter")
 include(":static-instrumenter:gradle-plugin")
